+++ src/utils/csdlParser.ts
@@ -1,14 +1,13 @@
-import { workspace } from "vscode";
+import { Uri, workspace } from "vscode";
 
 import { runOnlineParser } from './runOnlineParser';
 import { FieldWithType } from "../workbench/federationCompletionProvider";
-import { WorkbenchFileManager } from "../workbench/workbenchFileManager";
 
 export async function extractDefinedEntitiesByService() {
     let extendables: { [serviceName: string]: { type: string, keyFields: FieldWithType[] }[] } = {};
     try {
         let directivesState: { fields: Partial<FieldWithType>[], type?: string, serviceName?: string } = { fields: [] };
-        let textDoc = await workspace.openTextDocument(WorkbenchFileManager.composedSchemaPath);
+        let textDoc = await workspace.openTextDocument(Uri.parse('workspace:/csdl.graphql?csdl'));
 
         runOnlineParser(textDoc.getText(), (state, range, tokens) => {
             switch (state.kind) {